clean up URL handling
