+++ src/util/clientAPIInterface.js
@@ -11,7 +11,7 @@ export const getManifest = (dispatch, s3bucket = "live") => {
   };
   const processData = (data) => {
     const datasets = JSON.parse(data);
-    console.log("SERVER API REQUEST RETURNED:", datasets);
+    // console.log("SERVER API REQUEST RETURNED:", datasets);
     dispatch({
       type: MANIFEST_RECEIVED,
       s3bucket,
